// src/api/user/api.ts
export interface LoginRequest {
    /**
     * 
     * @type {AuthProvider}
     * @memberof LoginRequest
     */
    'provider': AuthProvider;
    /**
     * 
     * @type {UserCredentials}
     * @memberof LoginRequest
     */
    'credentials': UserCredentials;
}
// ...
export interface RegisterRequest {
    /**
     * 
     * @type {AuthProvider}
     * @memberof RegisterRequest
     */
    'provider': AuthProvider;
    /**
     * 
     * @type {UserRegistrationDetails}
     * @memberof RegisterRequest
     */
    'user': UserRegistrationDetails;
}
/**
 * 
 * @export
 * @interface ResendOtpRequest
 */
export interface ResendOtpRequest {
    /**
     * The user\'s email address to which the OTP will be resent.
     * @type {string}
     * @memberof ResendOtpRequest
     */
    'email': string;
}
/**
 * 
 * @export
 * @interface SendEmailOtpRequest
 */
export interface SendEmailOtpRequest {
    /**
     * The user\'s email to which the OTP will be sent.
     * @type {string}
     * @memberof SendEmailOtpRequest
     */
    'email': string;
}
// ...
export interface UserLogoUploadResponse {
    /**
     * The status of the call (e.g., success, error).
     * @type {string}
     * @memberof UserLogoUploadResponse
     */
    'status'?: string;
    /**
     * A list of messages related to the status of the request.
     * @type {Array<string>}
     * @memberof UserLogoUploadResponse
     */
    'messages'?: Array<string>;
    /**
     * 
     * @type {UserLogoUploadResponseData}
     * @memberof UserLogoUploadResponse
     */
    'data'?: UserLogoUploadResponseData;
}
/**
 * Contains relevant data. This could be user data for registration or tokens for login.
 * @export
 * @interface UserLogoUploadResponseData
 */
export interface UserLogoUploadResponseData {
    /**
     * The URL to access the uploaded logo.
     * @type {string}
     * @memberof UserLogoUploadResponseData
     */
    'url'?: string;
}
// ...
export interface UserUpdateDetails {
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'country'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'companyName'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'companyAddress'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'companyTaxNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'zipCode'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'site'?: string;
}
/**
 * 
 * @export
 * @interface VerifyEmailOtpRequest
 */
export interface VerifyEmailOtpRequest {
    /**
     * The user\'s email address.
     * @type {string}
     * @memberof VerifyEmailOtpRequest
     */
    'email': string;
    /**
     * The OTP code sent to the user\'s email.
     * @type {string}
     * @memberof VerifyEmailOtpRequest
     */
    'otp': string;
}
// ...
export interface UserApiLoginRequest {
    /**
     * 
     * @type {LoginRequest}
     * @memberof UserApiLogin
     */
    readonly loginRequest: LoginRequest
}

/**
 * Request parameters for registerUser operation in UserApi.
 * @export
 * @interface UserApiRegisterUserRequest
 */
export interface UserApiRegisterUserRequest {
    /**
     * 
     * @type {RegisterRequest}
     * @memberof UserApiRegisterUser
     */
    readonly registerRequest: RegisterRequest
}

/**
 * Request parameters for resendEmailOtp operation in UserApi.
 * @export
 * @interface UserApiResendEmailOtpRequest
 */
export interface UserApiResendEmailOtpRequest {
    /**
     * 
     * @type {ResendOtpRequest}
     * @memberof UserApiResendEmailOtp
     */
    readonly resendOtpRequest: ResendOtpRequest
}

/**
 * Request parameters for sendEmailOtp operation in UserApi.
 * @export
 * @interface UserApiSendEmailOtpRequest
 */
export interface UserApiSendEmailOtpRequest {
    /**
     * 
     * @type {SendEmailOtpRequest}
     * @memberof UserApiSendEmailOtp
     */
    readonly sendEmailOtpRequest: SendEmailOtpRequest
}

/**
 * Request parameters for updateCurrentUser operation in UserApi.
 * @export
 * @interface UserApiUpdateCurrentUserRequest
 */
export interface UserApiUpdateCurrentUserRequest {
    /**
     * 
     * @type {UserUpdateDetails}
     * @memberof UserApiUpdateCurrentUser
     */
    readonly userUpdateDetails: UserUpdateDetails
}

/**
 * Request parameters for uploadUserLogo operation in UserApi.
 * @export
 * @interface UserApiUploadUserLogoRequest
 */
export interface UserApiUploadUserLogoRequest {
    /**
     * The image file to be uploaded.
     * @type {File}
     * @memberof UserApiUploadUserLogo
     */
    readonly file?: File
}

/**
 * Request parameters for verifyEmailOtp operation in UserApi.
 * @export
 * @interface UserApiVerifyEmailOtpRequest
 */
export interface UserApiVerifyEmailOtpRequest {
    /**
     * 
     * @type {VerifyEmailOtpRequest}
     * @memberof UserApiVerifyEmailOtp
     */
    readonly verifyEmailOtpRequest: VerifyEmailOtpRequest
}

/**
 * UserApi - object-oriented interface
 * @export
 * @class UserApi
 * @extends {BaseAPI}
 */
export class UserApi extends BaseAPI {
    /**
     * Returns information about the authenticated user, including their roles and permissions.
     * @summary Get current user information
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public getCurrentUser(options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).getCurrentUser(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Logs in a user using credentials for a specific provider. Returns access and refresh tokens.
     * @summary Login with credentials from an authentication provider
     * @param {UserApiLoginRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public login(requestParameters: UserApiLoginRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).login(requestParameters.loginRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Registers a user with data provided from an authentication provider. If the user doesn\'t exist, creates a new account.
     * @summary Register a new user
     * @param {UserApiRegisterUserRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public registerUser(requestParameters: UserApiRegisterUserRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).registerUser(requestParameters.registerRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Resends a new OTP to the user\'s email address for verification.
     * @summary Resend OTP to user\'s email
     * @param {UserApiResendEmailOtpRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public resendEmailOtp(requestParameters: UserApiResendEmailOtpRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).resendEmailOtp(requestParameters.resendOtpRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Sends a one-time password (OTP) to the user\'s email address for verification purposes.
     * @summary Send OTP to email
     * @param {UserApiSendEmailOtpRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public sendEmailOtp(requestParameters: UserApiSendEmailOtpRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).sendEmailOtp(requestParameters.sendEmailOtpRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Allows the authenticated user to update their profile information, excluding the email.
     * @summary Update current user information
     * @param {UserApiUpdateCurrentUserRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public updateCurrentUser(requestParameters: UserApiUpdateCurrentUserRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).updateCurrentUser(requestParameters.userUpdateDetails, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Accepts an image file and responds with a URL to access the uploaded logo.
     * @summary Upload a user logo
     * @param {UserApiUploadUserLogoRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public uploadUserLogo(requestParameters: UserApiUploadUserLogoRequest = {}, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).uploadUserLogo(requestParameters.file, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Verifies the one-time password (OTP) submitted by the user for email confirmation.
     * @summary Verify OTP for email
     * @param {UserApiVerifyEmailOtpRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public verifyEmailOtp(requestParameters: UserApiVerifyEmailOtpRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).verifyEmailOtp(requestParameters.verifyEmailOtpRequest, options).then((request) => request(this.axios, this.basePath));
    }
}